Change the type of the entity ID from short to String
The Baba class now contains a String array simply listing
all the different entity names. (which are called "id"),
The mapping (id -> texture) is done in Rsrc,
readme of gui packages is updated.

This should help reduce confusion and increase speed and
easiness in creating a new type of entity.